refactor: mv placeholder license extractor to sources.local
